Creating an api folder to organize
