+++ Dockerfile
@@ -1,5 +1,4 @@
 FROM python:3.8
-#ENV PYTHONUNBUFFERRED 1
 
 COPY . /src
 WORKDIR /src
