+++ alembic/env.py
@@ -30,7 +30,7 @@ fileConfig(config.config_file_name)
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
 
-import database
+from api import database
 
 target_metadata = database.Base.metadata
